@manishucsd 2023-10-19
> There isn't a perfect solution to that problem. Tests need to compare against some reference. Golden test data is a pain to maintain. "Reference" implementations that are actually optimized are a combination of less reliable and more maintenance. And so on.

A few suggestions on speeding up reference runs / generating golden inputs. 

- Use a very simple device reference, where device reference is running on a GPU. It is a simplest implementation without using say SMEM (NVDA) or LDS (AMD).
- Use `torch.matmul` to run reference. It should probably be faster than codegen implementation. 
- Pre compute golden references on fixed random data and store hashcode. CUTLASS convolutions follow this approach as  device reference are also super slow. See example of [generating](https://github.com/NVIDIA/cutlass/blob/main/test/unit/conv/device/cache_testbed_output.h) and [using](https://github.com/NVIDIA/cutlass/tree/main/test/unit/data/hashes) hashcode